Added GCC_WEXTRA macro to determine if the C compiler  supports the -Wextra switch to enable extra warnings. Remove unused function parameters and address signed/unsigned comparisons highlighted by -Wextra.
git-svn-id: svn+ssh://svn.nta-monitor.com/trunk/opensource/arp-scan@18043 062a1500-4a13-0410-a63b-ee65f32af78f

+++ utils.c
@@ -172,7 +172,7 @@ make_message(const char *fmt, ...) {
       n = vsnprintf (p, size, fmt, ap);
       va_end(ap);
       /* If that worked, return the string. */
-      if (n > -1 && n < size)
+      if (n > -1 && n < (int) size)
          return p;
       /* Else try again with more space. */
       if (n > -1)    /* glibc 2.1 */
